added tests to test EquationSide splitting components correctly
