+++ src/main/java/microservice/equationSolver/controller/EquationSolverHandler.java
@@ -1,4 +1,4 @@
-package equationSolver.equationSolver.linearEquationSolver;
+package microservice.equationSolver.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
@@ -7,6 +7,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
+
+import microservice.equationSolver.linearEquationSolver.EquationHolder;
 
 /*
  *
